Function ssg::compiler::compile

pub fn compile(
    src_dir: &Path,
    out_dir: &Path,
    template_path: Option<&String>,
    site_name: String
) -> Result<(), Box<dyn Error>>

Compiles files in a source directory, generates HTML pages from them, and writes the resulting pages to an output directory. Also generates an index page containing links to the generated pages. This function takes in the paths to the source and output directories as arguments.

The function reads in all Markdown files in the source directory and extracts metadata from them. The metadata is used to generate appropriate meta tags for the resulting HTML pages. The Markdown content of the files is then converted to HTML using the generate_html function and rendered into complete HTML pages using the render_page function. The resulting pages are written to the output directory as HTML files with the same names as the original Markdown files.

Finally, the function generates an index HTML page containing links to all the generated pages. The resulting index page is written to the output directory as “index.html”.

If any errors occur during the process (e.g. a file cannot be read or written), an error is returned. Otherwise, Ok(()) is returned.

Arguments

src_dir - The path to the source directory.

out_dir - The path to the output directory.

template_path - The path to the template directory.

site_name - The name of the site.

Returns

Ok(()) if the compilation is successful. Err if an error occurs during the compilation. The error is wrapped in a Box<dyn Error> to allow for different error types.