Function ssg::template::create_template_folder

pub fn create_template_folder(
    template_path: Option<&String>
) -> Result<String, TemplateError>

Creates a template folder based on the provided template path or uses the default template folder

If a URL is provided as the template path, the function downloads the template files to a temporary directory.

If a local path is provided, the function uses it as the template directory path.

If no path is provided, the function downloads the default template files to a temporary directory.

Arguments

template_path - An optional &str containing the path to the template folder

Returns

Returns a Result that contains () if successful, or a TemplateError if an error occurs.